// src/app/shared-components/models/employee-models/employee-table-field-group.ts
import {FieldMetadataModel} from '../shared-models/field-metadata.model';

export class EmployeeTableFieldGroup {
  lastName: FieldMetadataModel;
  officeName: FieldMetadataModel;
  birthdate: FieldMetadataModel;
  directorLastName: FieldMetadataModel;
  directorId: FieldMetadataModel;
  directorFirstName: FieldMetadataModel;
  directorEmail: FieldMetadataModel;
  officeNameId: FieldMetadataModel;
  managerId: FieldMetadataModel;
  managerLastName: FieldMetadataModel;
  firstName: FieldMetadataModel;
  managerEmail: FieldMetadataModel;
  managerFirstName: FieldMetadataModel;
  id: FieldMetadataModel;
  someLabel: FieldMetadataModel;
  email: FieldMetadataModel;
}